force ecn option
without the "needs ecn" flag, flows don't always get ecn support
even if they are configured to. Not sure why. Don't have time
for tracking it down, so add a force ecn option in the meantime,
which will be used by the Mininet experiments to force use of ecn
or fallback functions (fallback doesn't work when "needs ecn" flag
is used).